requirements.txt
Django==5.1.7
djangorestframework==3.15.2
psycopg[binary]==3.2.6
python-dotenv==1.0.1
pytest==8.3.5
pytest-django==4.10.0
xhtml2pdf==0.2.17
